requirements.txt
streamlit==1.42.0
mecab-python3==1.0.10
pandas==2.2.3
python-docx==1.1.2
